+++ packages/database/src/entities/AppConfig.ts
@@ -1,6 +1,7 @@
+/* eslint-disable max-lines */
 /* istanbul ignore file */
 import { Entity, Property, EntityManager } from '@mikro-orm/core';
-import { ConstructorValues } from '../utils/types';
+import { ConstructorValues } from '../types/ConstructorValues';
 import { Node } from './Node';
 
 export enum ConfigKey {}
@@ -25,7 +26,7 @@ export class AppConfig extends Node<AppConfig> {
     this.value = value;
   }
 
-  getSafeKeys(): AppConfigPropertyKeys[] {
+  static getSafeKeys(): AppConfigPropertyKeys[] {
     return ['key', 'value'];
   }
