+++ packages/database/src/entities/JudgingVote/index.ts
@@ -1,0 +1,2 @@
+/* istanbul ignore file */
+export * from './JudgingVote'
