+++ packages/database/src/entities/Project/index.ts
@@ -1,0 +1,2 @@
+/* istanbul ignore file */
+export * from './Project'
