+++ packages/database/src/entities/index.ts
@@ -1,1 +1,9 @@
+/* istanbul ignore file */
 export * from './User';
+export * from './JudgingVote'
+export * from './Project'
+export * from './AppConfig'
+export * from './Event'
+export * from './Judge'
+export * from './Node'
+export * from './Prize'
